backend/requirements.txt
# Production Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
exa-py==1.0.9
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.3
certifi==2024.2.2
openai>=1.12.0
